Handle resource not found in Delete

+++ mongodb/src/main/java/de/idealo/shopping/mongodb/services/OrderService.java
@@ -1,9 +1,10 @@
 package de.idealo.shopping.mongodb.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
+import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -11,8 +12,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.client.HttpClientErrorException;
 
+//import de.idealo.shopping.mongodb.exceptions.ResourceNotFoundException;
 import de.idealo.shopping.mongodb.model.Order;
 import de.idealo.shopping.mongodb.repositories.OrderRepository;
 import de.idealo.shopping.mongodb.repositories.PersonRepository;
@@ -54,11 +55,11 @@ public class OrderService {
 
 	@DeleteMapping
 	public void delete(@RequestBody Order order) {
-		try {
-			orderRepository.delete(order);
-		} catch (IllegalStateException e) {
-			throw new HttpClientErrorException(HttpStatus.NOT_FOUND,
-					String.format("Order: {%s} cannot be found", order));
+		Optional<Order> dbOrder = orderRepository.findById(order.getId().toString());
+		if (!dbOrder.isPresent()) {
+			throw new ResourceNotFoundException(String.format("Order: {%s} cannot be found", order));
 		}
+		
+		orderRepository.delete(order);
 	}
 }
