+++ mongodb/src/main/java/de/idealo/shopping/mongodb/services/PersonService.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -17,6 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.HttpClientErrorException;
 
 import de.idealo.shopping.mongodb.model.Address;
+import de.idealo.shopping.mongodb.model.Order;
 import de.idealo.shopping.mongodb.model.Person;
 import de.idealo.shopping.mongodb.repositories.AddressRepository;
 import de.idealo.shopping.mongodb.repositories.PersonRepository;
@@ -71,11 +73,11 @@ public class PersonService {
 
 	@DeleteMapping
 	public void delete(@RequestBody Person person) {
-		try {
-			personRepository.delete(person);
-		} catch (IllegalStateException e) {
-			throw new HttpClientErrorException(HttpStatus.NOT_FOUND,
-					String.format("Person: {%s} cannot be found", person));
+		Optional<Person> dbPerson = personRepository.findById(person.getId().toString());
+		if (!dbPerson.isPresent()) {
+			throw new ResourceNotFoundException(String.format("Person: {%s} cannot be found", person));
 		}
+		
+		personRepository.delete(person);
 	}
 }
